+++ veltahleon/src/main/java/com/testament/veltahleon/services/implementation/society/PersonServiceImpl.java
@@ -1,8 +1,8 @@
 package com.testament.veltahleon.services.implementation.society;
 
 
+import com.testament.veltahleon.mapper.society.PersonMapper;
 import com.testament.veltahleon.model.entities.history.Race;
-import com.testament.veltahleon.model.entities.society.Family;
 import com.testament.veltahleon.model.entities.society.Person;
 import com.testament.veltahleon.repositories.society.PersonRepository;
 import com.testament.veltahleon.services.ifc.society.PersonService;